requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-multipart
pytesseract
Pillow
python-jose[cryptography]
passlib[bcrypt]
alembic
python-dateutil
